cleaned up token hooks
