+++ packages/web-app/src/hooks/useDaoBalances.tsx
@@ -1,19 +1,8 @@
 import {Address} from '@aragon/ui-components/dist/utils/addresses';
 import {useLazyQuery} from '@apollo/client';
 
+import {DaoTokenBalance} from 'utils/types';
 import {DAO_BALANCE_LIST} from 'queries/finances';
-
-type BalanceFromGraph = {
-  id: number;
-  token: {
-    id: string;
-    name: string;
-    symbol: string;
-    decimals: number;
-  };
-  balance: bigint;
-  lastUpdated: string;
-};
 
 export const useDaoBalances = (daoAddress: Address) => {
   const [getDaoBalances, {data, error, loading, refetch}] = useLazyQuery(
@@ -22,7 +11,7 @@ export const useDaoBalances = (daoAddress: Address) => {
   );
 
   return {
-    data: data?.balances as BalanceFromGraph[],
+    data: data?.balances as DaoTokenBalance[],
     error,
     loading,
     refetch,
