+++ packages/web-app/src/queries/tokenData.tsx
@@ -1,0 +1,16 @@
+import {gql} from '@apollo/client';
+
+export const TOKEN_DATA_QUERY = gql`
+  query TokenData {
+    tokenData(url: $url)
+      @rest(type: "TokenData", path: "{args.url}", method: "GET") {
+      id
+      name
+      symbol
+      image {
+        large
+      }
+      address
+    }
+  }
+`;
